Move Resource into asset package
